requirements.txt
-i https://pypi.org/simple
accelerate==1.1.0; python_full_version >= '3.9.0'
aiofiles==23.2.1; python_version >= '3.7'
aiohappyeyeballs==2.4.3; python_version >= '3.8'
aiohttp==3.10.10; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
airportsdata==20241001; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anthropic==0.39.0; python_version >= '3.8'
anyio==4.6.2.post1; python_version >= '3.9'
attrs==24.2.0; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cloudpickle==3.1.0; python_version >= '3.8'
datasets==3.1.0; python_full_version >= '3.8.0'
dill==0.3.8; python_version >= '3.8'
diskcache==5.6.3; python_version >= '3'
distro==1.9.0; python_version >= '3.6'
docstring-parser==0.16; python_version >= '3.6' and python_version < '4.0'
fastapi==0.115.4; python_version >= '3.8'
ffmpy==0.4.0; python_full_version >= '3.8.1' and python_full_version < '4.0.0'
filelock==3.16.1; python_version >= '3.8'
frozenlist==1.5.0; python_version >= '3.8'
fsspec[http]==2024.9.0; python_version >= '3.8'
gradio==5.5.0; python_version >= '3.10'
gradio-client==1.4.2; python_version >= '3.10'
greenlet==3.1.1; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.6; python_version >= '3.8'
httpx==0.27.2; python_version >= '3.8'
huggingface-hub==0.26.2; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
iniconfig==2.0.0; python_version >= '3.7'
instructor==1.6.3; python_version >= '3.9' and python_version < '4.0'
interegular==0.3.3; python_version >= '3.7'
jinja2==3.1.4; python_version >= '3.7'
jiter==0.5.0; python_version >= '3.8'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
lark==1.2.2; python_version >= '3.8'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
mdurl==0.1.2; python_version >= '3.7'
mpmath==1.3.0
multidict==6.1.0; python_version >= '3.8'
multiprocess==0.70.16; python_version >= '3.8'
nest-asyncio==1.6.0; python_version >= '3.5'
networkx==3.4.2; python_version >= '3.10'
numpy==1.26.4; python_version >= '3.9'
openai==1.54.0; python_version >= '3.8'
orjson==3.10.11; python_version >= '3.8'
outlines==0.1.1; python_version >= '3.9'
outlines-core==0.1.14; python_version >= '3.8'
packaging==24.1; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pillow==11.0.0; python_version >= '3.9'
playwright==1.48.0; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
propcache==0.2.0; python_version >= '3.8'
psutil==6.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyarrow==18.0.0; python_version >= '3.9'
pycountry==24.6.1; python_version >= '3.8'
pydantic==2.9.2; python_version >= '3.8'
pydantic-core==2.23.4; python_version >= '3.8'
pydub==0.25.1
pyee==12.0.0; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pytest==8.3.3; python_version >= '3.8'
pytest-base-url==2.1.0; python_version >= '3.8'
pytest-playwright==0.5.2; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
python-multipart==0.0.12; python_version >= '3.8'
python-slugify==8.0.4; python_version >= '3.7'
pytz==2024.2
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.35.1; python_version >= '3.8'
regex==2024.9.11; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
rich==13.9.4; python_full_version >= '3.8.0'
rpds-py==0.20.1; python_version >= '3.8'
ruff==0.7.2; sys_platform != 'emscripten'
safehttpx==0.1.1; python_version >= '3.9'
safetensors==0.4.5; python_version >= '3.7'
semantic-version==2.10.0; python_version >= '2.7'
setuptools==75.3.0; python_version >= '3.12'
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
starlette==0.41.2; sys_platform != 'emscripten'
sympy==1.13.1; python_version >= '3.9'
tenacity==9.0.0; python_version >= '3.8'
text-unidecode==1.3
tokenizers==0.20.2; python_version >= '3.7'
tomlkit==0.12.0; python_version >= '3.7'
torch==2.5.1; python_full_version >= '3.8.0'
tqdm==4.66.6; python_version >= '3.7'
transformers==4.46.1; python_full_version >= '3.8.0'
typer==0.12.5; sys_platform != 'emscripten'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2024.2; python_version >= '2'
urllib3==2.2.3; python_version >= '3.8'
uvicorn==0.32.0; sys_platform != 'emscripten'
websockets==12.0; python_version >= '3.8'
xxhash==3.5.0; python_version >= '3.7'
yarl==1.17.1; python_version >= '3.9'
